Extracted history purger logic from PurgeHistoryResponder.
